+++ src/autoanki/BookCleaner/BookCleaner.py
@@ -1,27 +1,22 @@
 import logging
 import os
 
-# TODO: I think that the entire book cleaner file is not needed. 
+# TODO: I think that the entire book cleaner file is not needed.
 #   Words should not be written back to a file. This is slow
-CLEANED_FILES_DIRECTORY = 'cleaned_files'
-CLEANED_FILES_SUFFIX = '_cleaned'
+CLEANED_FILES_DIRECTORY = "cleaned_files"
+CLEANED_FILES_SUFFIX = "_cleaned"
 # Sentences That should not be added to the completed file.
-GARBAGE_SENTENCES = ['',
-                     "",
-                     ".",
-                     "。",
-                     "\n"]
+GARBAGE_SENTENCES = ["", "", ".", "。", "\n"]
 
 
 class BookCleaner:
-
     def __init__(self, debug_level, force=False):
-        """ Internal tool used to sanatize input
+        """Internal tool used to sanatize input
         Use `clean(bookpath)` to sanatize files and remove junk data
         Args:
             `force`: Ignore confirmations on if you want to clean >50 files
         """
-        self.logger = logging.getLogger('autoanki.bookcleaner')
+        self.logger = logging.getLogger("autoanki.bookcleaner")
         self.logger.setLevel(debug_level)
         self.file_list = []
         self.bookpath = ""
@@ -30,9 +25,9 @@ class BookCleaner:
     def clean(self, bookpath: str) -> None | list[str]:
         """
         Cleans the files contained in the bookpath. If bookpath is a single file, clean it.
-        Args: 
+        Args:
             `bookpath: filepath of files to be cleaned`
-        Return: 
+        Return:
             `str: list of cleaned file(s)`
         """
         if not os.path.exists(bookpath):
@@ -51,13 +46,15 @@ class BookCleaner:
                     # Only clean files that are not in cleaned_files directory
                     in_cleaned = str(root).find(CLEANED_FILES_DIRECTORY) != -1
                     if not in_cleaned:
-                        if '.txt' in file:
+                        if ".txt" in file:
                             dirty_files.append(os.path.join(root, file))
 
             # Check this cleaning won't be mean to the CPU
             if len(dirty_files) > 50 and not self.force:
-                yn = input("Over 50 files. Are you sure you want to convert this many files? (Y/N)")
-                if yn.lower() != 'y':
+                yn = input(
+                    "Over 50 files. Are you sure you want to convert this many files? (Y/N)"
+                )
+                if yn.lower() != "y":
                     return None
 
         # Now we have a list of files to convert in a list
@@ -69,7 +66,9 @@ class BookCleaner:
             os.mkdir(cleaned_files_root)
 
         for file in dirty_files:
-            cleaned_filepath = self._clean_file(file, cleaned_files_root=cleaned_files_root)
+            cleaned_filepath = self._clean_file(
+                file, cleaned_files_root=cleaned_files_root
+            )
             if cleaned_filepath:
                 cleaned_files.append(cleaned_filepath)
 
@@ -90,30 +89,30 @@ class BookCleaner:
         self.logger.debug("New filepath: " + new_filepath)
 
         # Clean page file of characters that may cause issues.
-        page_file = open(filepath, encoding='utf-8')
-        try: 
+        page_file = open(filepath, encoding="utf-8")
+        try:
             page_sentences = page_file.read().split("。")
         except:
             self.logger.error(f"Critical error cleaning file: [{filepath}]")
             return
-        cleaned_file = open(new_filepath, "w", encoding='utf-8')
+        cleaned_file = open(new_filepath, "w", encoding="utf-8")
 
         for page_sentence in page_sentences:
             # Clean string
-            page_sentence = page_sentence\
-                .lstrip()\
-                .rstrip()\
-                .replace("  ", " ")\
-                .lstrip("\"")\
-                .rstrip("\"")\
-                .replace("”","'")\
-                .replace("　", "")\
-                .replace("“","")\
-                .rstrip("。")\
-                .strip("'")\
+            page_sentence = (
+                page_sentence.lstrip()
+                .rstrip()
+                .replace("  ", " ")
+                .lstrip('"')
+                .rstrip('"')
+                .replace("”", "'")
+                .replace("　", "")
+                .replace("“", "")
+                .rstrip("。")
                 .strip("'")
+                .strip("'")
+            )
             if page_sentence not in GARBAGE_SENTENCES:
                 cleaned_file.write(page_sentence + "。" + "\n")
 
         return new_filepath
-