+++ src/autoanki/Tokenizer/__init__.py
@@ -1,1 +1,1 @@
-from .ChineseTokenizer import ChineseTokenizer 
+from .ChineseTokenizer import ChineseTokenizer