+++ tests/test_DeckManager.py
@@ -3,18 +3,18 @@ from autoanki.AutoAnki import AutoAnki
 
 # @unittest.skip('Some reason')
 
+
 class TestDeckManager(unittest.TestCase):
+    @classmethod
+    def setUpClass(cls):  ### run once before all test cases ###
+        pass
 
     @classmethod
-    def setUpClass(cls): ### run once before all test cases ###
+    def tearDownClass(cls):  ### run once after all test cases ###
         pass
 
-    @classmethod
-    def tearDownClass(cls): ### run once after all test cases ###
+    def setUp(self):  ### run before each test case ###
         pass
 
-    def setUp(self): ### run before each test case ###
-        pass
-
-    def tearDown(self): ### run after each test case ###
+    def tearDown(self):  ### run after each test case ###
         pass